requirements.txt
aiohttp==3.8.6
aiohttp-retry==2.8.3
aiosignal==1.3.1
amqp==5.2.0
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
billiard==4.2.0
boto3==1.29.7
botocore==1.32.7
celery==5.3.6
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
Django==4.2.7
django-cors-headers==4.3.0
django-multiselectfield==0.1.12
django-restframework==0.0.1
django-storages==1.14.2
djangorestframework==3.14.0
docopt==0.6.2
frozenlist==1.4.0
idna==3.4
jmespath==1.0.1
kombu==5.3.4
multidict==6.0.4
Pillow==10.1.0
pipreqs==0.4.13
prompt-toolkit==3.0.41
PyJWT==2.8.0
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3.post1
redis==5.0.1
requests==2.31.0
s3transfer==0.8.0
six==1.16.0
sqlparse==0.4.4
twilio==8.10.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.7
vine==5.1.0
wcwidth==0.2.12
yarg==0.1.9
yarl==1.9.2
